+++ CHANGELOG.md
@@ -1,7 +1,7 @@
-### NEXT_VERSION_TYPE=MINOR
-### NEXT_VERSION_DESCRIPTION_BEGIN
+## [7.2.0](https://github.com/yoomoney-gradle-plugins/check-dependencies-plugin/pull/5) (19-03-2021)
+
 * Сборка проекта переведена на gradle-project-plugin.
-### NEXT_VERSION_DESCRIPTION_END
+
 ## [7.1.2](https://github.com/yoomoney-gradle-plugins/check-dependencies-plugin/pull/4) (11-03-2021)
 
 * Исправлен баг в ArtifactVersionResolver - теперь при поиске в списке репозиториев пропускаем репозитории, получение версий из